requirements.txt
Flask
gunicorn
google-generativeai
google-genai
python-dotenv


